changes server Client.send to accept (url, data) and send same JSON format as the clientside Server module ({url: url, data: data})
I don't think this is optimal - the old way was nice as there were no keys to maintain
aside from the urls - but it fits much better with how the server, client and publishers
are actually used (sending data for one url at a time and being able to build the url
inline).  [url, data] would be shorter/no keys to maintain but more obfuscated.