adds ClientConnected event to User and removes sendCurrentGames
ClientConnected is fired when the user opens a new connection (e.g.
when they open a new window) - so that initialisation messages can
be sent to each new client.

the app doesn't bother giving the user the full game list any more
to see whether they are in any games - if the session gets garbage
collected, they lose their games.

NOTE nothing is done yet about actually maintaining the list of games
in the session...