removes FIXME about game restoration needing to know original player details from Application.
Users can't log in or out without finishing all their current games, so there
will never be an interrupted game where a player's details changed.

To clarify, there can be games where a player logs out after the game is over,
so the details of the User won't be the same as when the game started, but that's
another issue and not really important as far as the code is concerned.  It's
more that the user will want to see the original player usernames when they look
at a finished game.