outlines the full procedure for game restoration and adds most of the implementation
Game still needs to do something with the game restoration details it's
given as a new argument, and it needs to handle its own clock and have
its own _startTime property so that it can make the necessary adjustments
for restored games.

Clock will also need a method for arbitrarily adding time to the active
player's timeLeft.