removes estimated crash time code from Application._submitGameRestorationRequest.
The best way of making sure the active player has time is to simply make it
seem as though the last move has only just been played.  This might result in
the active player getting a time advantage in some cases, but in time-sensitive
games the difference between that and estimating it accurately is likely to be
small anyway, and this way removes the complication and possibility of bigger
errors that would come with estimating it.  (the other problem with that is
that one or both of the players could be offline when it crashes, in which case
they won't know when it crashed at all).