starts Player class, starts moving send and subscribe code out of Game and into User (using snippets.js as a holding area).
Game will have public methods and events instead of interacting with
Users through the send/subscribe system.  It now deals with Players as
opposed to Users, and both Users and Bots will have a Player for playing
games.

Ratings will be handled by Users and Bots - Player just has an integer
rating with a setter and a getter.

Now that the methods are public, there is less control over who can do
what to a Game - even though you pass a Player, a User could easily call
game.move(game.getPlayer(game.getPlayerColour(this._player).opposite), ...
to make a move on behalf of their opponent.  This isn't a security issue,
it just means the security concerns are in the User now as opposed to in
the Game.