changes Player to be an adapter between Users and the interface that Game etc use (just getRating), and to handle User replacement by having a setUser method.
The Game doesn't have to handle Replaced because the Player stays the same, but
the Player's reference to a User changes, as do the Users' internal Players -
the new User gets the old User's Player, and the old User creates a new Player
(this happens on logout).

The /user/replaced functionality and the Replaced event are now gone; the user
just logs out when they're replaced.