Revert "refactors Bot to keep stockfish running during a game, increases movetime and removes setTimeout delay from move function"
This reverts commit bdcbaf6647ead9caec226f5c3ca3a74fd2daa084.

it didn't work - there seems to be something wrong or not understood properly
about how node communicates with other processes.  entering the same input on
the command line as Bot does with stdin.write() works normally, but for some
reason the bots will only make one move - after that the "data" handler for
stdout doesn't fire so either the input or output stops working.