implements player connection status feeds and isUser() methods on Player and Bot
this allows you to check whether a player-like object (Player or Bot)
represents a bot or a human user.

feeds of /player_connection_status/id are set up for spectators and players,
for all relevant players who are human users.  the client just needs to
handle the appropriate messages to get updates when a player d/cs or
reconnects (e.g. /player_connection_status/123).  the message is true if
connected and false if disconnected.