initial commit - changes Game.toJSON not to send history by default
this is to test the performance characteristics and ui-feel of loading just
the game details initially, and then requesting the moves and having them
sent to the client one-by-one.

unlike the last attempt (game_json_without_history) it doesn't have the
disadvantage of having to do heavy processing once moves arrive on the
client anyway, because of the performance improvements in the 2.0.x chess
and jsonchess series.